added proper error handling

+++ internal/services/user_service.go
@@ -2,10 +2,9 @@ package services
 
 import (
 	"context"
-	"log"
+	"users-service/internal/errors"
 	"users-service/internal/models"
 	"users-service/internal/repository"
-	"users-service/pkg/utils"
 
 	"github.com/google/uuid"
 )
@@ -22,9 +21,17 @@ func (s *UserService) CreateUser(ctx context.Context, input models.CreateUserInp
 
 	id := uuid.New().String()
 
-	passwordHash, err := utils.HashPassword(input.Password)
-	if err != nil {
-		return nil, err
+	// passwordHash, err := utils.HashPassword(input.Password)
+	// if err != nil {
+	// 	return nil, err
+	// }
+
+	if input.Email == "" {
+		return nil, errors.New(errors.CodeValidation, "email is required")
+	}
+
+	if input.Name == "" {
+		return nil, errors.New(errors.CodeValidation, "name is required")
 	}
 
 	user := models.User{
@@ -32,10 +39,10 @@ func (s *UserService) CreateUser(ctx context.Context, input models.CreateUserInp
 		Email:  input.Email,
 		Name:   input.Name,
 		Role:   "USER",
-		Active: 1,
+		Status: 1,
 	}
 
-	err = s.Repo.Create(ctx, user, passwordHash)
+	err := s.Repo.Create(ctx, user)
 	if err != nil {
 		return nil, err
 	}
@@ -50,7 +57,11 @@ func (s *UserService) GetAllUsers(ctx context.Context) ([]*models.User, error) {
 }
 
 func (s *UserService) GetUserById(ctx context.Context, id string) (*models.User, error) {
-	user, err := s.Repo.GetUser(ctx, id)
+
+	if id == "" {
+		return nil, errors.New(errors.CodeValidation, "id is required")
+	}
+	user, err := s.Repo.GetByID(ctx, id)
 
 	if err != nil {
 		return nil, err
@@ -59,16 +70,32 @@ func (s *UserService) GetUserById(ctx context.Context, id string) (*models.User,
 }
 
 func (s *UserService) UpdateUserById(ctx context.Context, id string, params models.UpdateUserDetails) error {
+
+	if id == "" {
+		return errors.New(errors.CodeValidation, "id is required")
+	}
+
+	// check if there is even one value that needs to be change
+	if params.Name == nil && params.Email == nil {
+		return errors.New(
+			errors.CodeValidation,
+			"at least one field must be updated",
+		)
+	}
+
 	err := s.Repo.UpdateUser(ctx, id, params)
 
 	if err != nil {
-		log.Println(err)
 		return err
 	}
 	return nil
 }
 
 func (s *UserService) DeleteUser(ctx context.Context, id string) error {
+
+	if id == "" {
+		return errors.New(errors.CodeValidation, "id is required")
+	}
 	if err := s.Repo.DeleteUser(ctx, id); err != nil {
 		return err
 	}
@@ -76,12 +103,16 @@ func (s *UserService) DeleteUser(ctx context.Context, id string) error {
 }
 
 func (s *UserService) GetUserByEmail(ctx context.Context, email string) (*models.User, error) {
+
+	if email == "" {
+		return nil, errors.New(errors.CodeValidation, "email is required")
+	}
 	user, err := s.Repo.GetUserByEmail(ctx, email)
 
 	if err != nil {
 		return nil, err
 	}
 
-	return user, err
+	return user, nil
 
 }
